+++ typescript/radix-sdk/src/clients/protocol.ts
@@ -125,6 +125,7 @@ export class RadixProtocolProvider implements ProtocolProvider {
       TEST_SEND_GAS: 0n,
       AVS_GAS: 0n,
       ISM_DEPLOY_GAS: 0n,
+      HOOK_DEPLOY_GAS: 0n,
     };
   }